[refactor] : 체험 상세 디자인 수정
