--- utile/initialDevice.tsx
@@ -1,11 +1,0 @@
-const initialDevice = () => {
-  let device = 'mobile';
-  if (window.innerWidth > 1199) {
-    device = 'windows';
-  } else if (window.innerWidth > 767) {
-    device = 'tablet';
-  }
-  return device;
-};
-
-export {initialDevice};
